move submission processor to its own class
